Completely repair real type error

+++ presto-base-elasticsearch/src/main/java/com/facebook/presto/elasticsearch/EsTypeManager.java
@@ -20,7 +20,6 @@ import static com.facebook.presto.spi.type.BigintType.BIGINT;
 import static com.facebook.presto.spi.type.BooleanType.BOOLEAN;
 import static com.facebook.presto.spi.type.DoubleType.DOUBLE;
 import static com.facebook.presto.spi.type.IntegerType.INTEGER;
-import static com.facebook.presto.spi.type.RealType.REAL;
 import static com.facebook.presto.spi.type.SmallintType.SMALLINT;
 import static com.facebook.presto.spi.type.TimestampType.TIMESTAMP;
 import static com.facebook.presto.spi.type.VarbinaryType.VARBINARY;
@@ -82,7 +81,7 @@ public class EsTypeManager
                 type = INTEGER;
                 break;
             case FLOAT:
-                type = REAL;
+                type = DOUBLE;
                 break;
             case SCALED_FLOAT:
             case HALF_FLOAT:
